Add a script that scrapes Wikipedia's truck manufacturer's page and puts it in a csv file
get_truck_brand_names.py, when run, will scrape Wikipedia's List of truck manufacturers page, parse it, handle most of the weird entries,
and return the truck_brands.csv file which can then be used for different purposes.
Two things are of note. First, the script doesn't handle all the errors on Wikipedia page. Especially the country of origin field is messy
and in need of additional processing. Second, the script needs to be run only once outside the main ChatBot functionality, and is not
called by the ChatBot itself. The ChatBot script will, however, use the csv file for comparison of brand names.