Fix a save&quit bug
Issue: if a user typed 'q'/'quit' in the very first line, there
wouldn't be enough information (namely, username and fleet_id) to
construct the paths for output data, and the conversation file
would get a fixed name based on the last dialogue line. Repeating
this would then break the script, as the program tries to create
a second conversation file with preciesly the same name (due to
`x` flag when opening output data file).

Expected behavior: no matter when the user quits, output files
should be created normally and in a desired location.

Reproduce: try to quit the script right at the start of the chat,
two times in a row.

Solution: a very quick and dirty patch that creates a random
username and fleet ID, and uses them to initialize the output
data files.